mod : MNIST preprocessing - Chunmyong

+++ MNIST_Tutorial/MLP_MSE_With_MNIST.cpp
@@ -55,7 +55,9 @@ int main(int argc, char const *argv[]) {
 
         HGUNN.Run(optimizer);
 
-        if ((i % 100) == 0) std::cout << "Train Accuracy is : " << temp::Accuracy(act2->GetOutput(), label->GetOutput(), BATCH) << '\n';
+        if ((i % 100) == 0) std::cout << "Train Accuracy is : "
+                                      << (float)temp::Accuracy(act2->GetOutput(), label->GetOutput(), BATCH)
+                                      << '\n';
     }
 
     // ======================= Testing =======================
@@ -68,7 +70,7 @@ int main(int argc, char const *argv[]) {
 
         HGUNN.Run(err);
         // I'll implement flexibility about the situation that change of Batch size
-        test_accuracy += temp::Accuracy(act2->GetOutput(), label->GetOutput(), BATCH);
+        test_accuracy += (float)temp::Accuracy(act2->GetOutput(), label->GetOutput(), BATCH);
     }
 
     std::cout << "Test Accuracy is : " << test_accuracy / (int)LOOP_FOR_TEST << '\n';
